init, intexpr, list, memmgt
